Move gui tooltip formatting to own file
